feat(scheduler, blobnode): add switch to support shard partial repair
1. Unified error return for shard partial repair

Closes #2377

+++ blobstore/api/blobnode/iostat.go
@@ -46,7 +46,7 @@ func (it IOType) IsValid() bool {
 }
 
 func (it IOType) String() string {
-	return IOtypemap[uint64(it)]
+	return IOtypemap[it]
 }
 
 func (it IOType) IsHighLevel() bool {
