+++ blobstore/blobnode/service.go
@@ -72,7 +72,7 @@ func initPromeConf() {
 		metricCfg.Service = proto.ServiceNameBlobNode
 	}
 	if len(metricCfg.SizeBuckets) == 0 {
-		metricCfg.SizeBuckets = []int{0, 65536, 131072, 262144, 524288, 1048576, 4194304, 8388608, 16777216, 33554432}
+		metricCfg.SizeBuckets = []int64{0, 65536, 131072, 262144, 524288, 1048576, 4194304, 8388608, 16777216, 33554432}
 	}
 	if metricCfg.SetDefaultSwitch {
 		metricCfg.EnableHttpMethod = true
